check point -> masih error di UserService dan RegistrationService

+++ src/main/java/mii/mcc72/ams_server_app/repos/ConfirmationTokenRepository.java
@@ -20,7 +20,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Repository
 @Transactional(readOnly = true)
 public interface ConfirmationTokenRepository
-        extends JpaRepository<ConfirmationToken, Long> {
+        extends JpaRepository<ConfirmationToken, Integer> {
 
     Optional<ConfirmationToken> findByToken(String token);
